corpusv1/requirements.txt
requests==2.28.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.3.1
bs4==0.0.2